Show confirmation modal when adding the project
