+++ django_project/base/views/__init__.py
@@ -1,2 +1,3 @@
 from project import *  # noqa
 from error_views import *  # noqa
+from github_project import *  # noqa
